hello triangle

+++ rebuild.sh
@@ -2,8 +2,11 @@
 
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 
-rm -rf "$SCRIPT_DIR/build"
-mkdir "$SCRIPT_DIR/build"
+if [ "$1" = "all" ]; then
+    echo "From scratch"
+    rm -rf "$SCRIPT_DIR/build"
+    mkdir "$SCRIPT_DIR/build"
+fi
 pushd "$SCRIPT_DIR/build"
 cmake ..
 make -j4
